Added CmsTag and acts_as_taggable class and associated tests
CmsTag allows module and core developers to globally "tag" content
(taxonomy without the hierarchy) for web 2.0-style tagging of content
items.  This will come into play later on in the 2.0 development cycle
when we start to examine ways to allow content to not be quite so
structured in all cases.

git-svn-id: http://svn.cmsmadesimple.org/svn/cmsmadesimple/trunk@5034 3d254a34-79dc-0310-9e5f-be208747d8a0